requirements.txt
Flask>=0.10.1
Flask-WTF
Flask-Script
celery
pygithub
redis
